Fix SubWithBorrow sign-extension bug on 32-bit and update stale comments
SubWithBorrow's 32-bit path cast nuint operands through (int) before
widening to long, which sign-extends values >= 0x80000000 instead of
zero-extending. For example, SubWithBorrow(0xFFFFFFFF, 0, 0) would
incorrectly return borrowOut=1 instead of 0.

Fix: cast directly to long (zero-extension for unsigned nuint/uint).

Also update FastReducer comments from '2^(32*k)' to '2^(kcbitNuint*k)'
to reflect the variable limb width.

+++ src/libraries/System.Runtime.Numerics/src/System/Numerics/BigIntegerCalculator.Utils.cs
@@ -131,7 +131,7 @@ namespace System.Numerics
             }
             else
             {
-                long diff = (long)(int)a - (int)b - (int)borrowIn;
+                long diff = (long)a - (long)b - (long)borrowIn;
                 borrowOut = (nuint)(uint)(-(int)(diff >> 32)); // 0 or 1
                 return (nuint)(uint)diff;
             }
